[PackageManager] Separate PackageManager from policies
Moves `PackageManager` out from `sos/policies/__init__.py` into a new
`sos/policies/package_managers` subdir.

Future commits will aim to canonicalize package manager subclasses for
policies to use, and ease the creation of new reusable package managers.

Related: #2349
