+++ sos/policies/distros/suse.py
@@ -12,7 +12,7 @@ import sys
 
 from sos.report.plugins import RedHatPlugin, SuSEPlugin
 from sos.policies.distros import LinuxPolicy
-from sos.policies import PackageManager
+from sos.policies.package_managers import PackageManager
 from sos import _sos as _
